Add files via upload
React (also known as React. js or ReactJS) is a free and open-source front-end JavaScript library for building user interfaces based on components.
The definition is a simple one. It's a list of tasks you need to complete or things that you want to do. Most typically, they're organised in order of priority. Traditionally, they're written on a piece of paper or post it notes and act as a memory aid.
here I have used useState and useEffect events in this React Todolist project.